Refactoring - isolated the certificate authentication logic to static helper so that it's accessible in isolation
