+++ src/Stratiteq.Microservices.ClientCertAuth/CertificateAuthenticationHandler.cs
@@ -1,22 +1,15 @@
 ﻿// Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-using System;
-using System.Security.Claims;
-using System.Security.Cryptography;
-using System.Security.Cryptography.X509Certificates;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Stratiteq.Microservices.X509Certificate;
 
 namespace Stratiteq.Microservices.ClientCertAuth
 {
     internal class CertificateAuthenticationHandler : AuthenticationHandler<CertificateAuthenticationOptions>
     {
-        private static readonly Oid ClientCertificateOid = new Oid("1.3.6.1.5.5.7.3.2");
-
         public CertificateAuthenticationHandler(
             IOptionsMonitor<CertificateAuthenticationOptions> options,
             ILoggerFactory logger,
@@ -42,101 +35,13 @@ namespace Stratiteq.Microservices.ClientCertAuth
         /// <returns>A new instance of the events instance.</returns>
         protected override Task<object> CreateEventsAsync() => Task.FromResult<object>(new CertificateAuthenticationEvents());
 
-        protected override async Task<AuthenticateResult> HandleAuthenticateAsync()
-        {
-            // You only get client certificates over HTTPS
-            if (!Context.Request.IsHttps)
-            {
-                return AuthenticateResult.NoResult();
-            }
-
-            var clientCertificate = await Context.Connection.GetClientCertificateAsync();
-
-            // This should never be the case, as cert authentication happens long before ASP.NET kicks in.
-            if (clientCertificate == null)
-            {
-                Logger.LogDebug("No client certificate found.");
-                return AuthenticateResult.NoResult();
-            }
-
-            // If we have a self signed cert, and they're not allowed, exit early and not bother with
-            // any other validations.
-            if (clientCertificate.IsSelfSigned() &&
-                !Options.AllowedCertificateTypes.HasFlag(CertificateTypes.SelfSigned))
-            {
-                Logger.LogWarning("Self signed certificate rejected, subject was {0}", clientCertificate.Subject);
-
-                return AuthenticateResult.Fail("Options do not allow self signed certificates.");
-            }
-
-            // If we have a chained cert, and they're not allowed, exit early and not bother with
-            // any other validations.
-            if (!clientCertificate.IsSelfSigned() &&
-                !Options.AllowedCertificateTypes.HasFlag(CertificateTypes.Chained))
-            {
-                Logger.LogWarning("Chained certificate rejected, subject was {0}", clientCertificate.Subject);
-
-                return AuthenticateResult.Fail("Options do not allow chained certificates.");
-            }
-
-            var chainPolicy = BuildChainPolicy(clientCertificate);
-
-            try
-            {
-                var chain = new X509Chain { ChainPolicy = chainPolicy };
-
-                var certificateIsValid = chain.Build(clientCertificate);
-
-                if (!certificateIsValid)
-                {
-                    using (Logger.BeginScope(clientCertificate.SHA256Thumprint()))
-                    {
-                        Logger.LogWarning("Client certificate failed validation, subject was {0}", clientCertificate.Subject);
-                        foreach (var validationFailure in chain.ChainStatus)
-                        {
-                            Logger.LogWarning("{0} {1}", validationFailure.Status, validationFailure.StatusInformation);
-                        }
-                    }
-
-                    return AuthenticateResult.Fail("Client certificate failed validation.");
-                }
-
-                var validateCertificateContext = new ValidateCertificateContext(Context, Scheme, Options) { ClientCertificate = clientCertificate };
-
-                await Events.ValidateCertificate(validateCertificateContext);
-
-                if (validateCertificateContext.Result != null &&
-                    validateCertificateContext.Result.Succeeded)
-                {
-                    return Success(validateCertificateContext.Principal, clientCertificate);
-                }
-
-                if (validateCertificateContext.Result != null &&
-                    validateCertificateContext.Result.Failure != null)
-                {
-                    return AuthenticateResult.Fail(validateCertificateContext.Result.Failure);
-                }
-
-                var certClaims = clientCertificate.CreateClaimsFromCertificate(Options.ClaimsIssuer);
-                return Success(new ClaimsPrincipal(new ClaimsIdentity(certClaims, CertificateAuthenticationDefaults.AuthenticationScheme)), clientCertificate);
-            }
-            catch (Exception ex)
-            {
-                var authenticationFailedContext = new CertificateAuthenticationFailedContext(Context, Scheme, Options)
-                {
-                    Exception = ex,
-                };
-
-                await Events.AuthenticationFailed(authenticationFailedContext);
-
-                if (authenticationFailedContext.Result != null)
-                {
-                    return authenticationFailedContext.Result;
-                }
-
-                throw;
-            }
-        }
+        protected override async Task<AuthenticateResult> HandleAuthenticateAsync() =>
+            await CertificateAuthenticationHelper.AuthenticateCertificateAsync(
+                Context,
+                Logger,
+                Options,
+                Events,
+                Scheme);
 
         protected override Task HandleChallengeAsync(AuthenticationProperties properties)
         {
@@ -150,59 +55,6 @@ namespace Stratiteq.Microservices.ClientCertAuth
         {
             Response.StatusCode = 403;
             return Task.CompletedTask;
-        }
-
-        private X509ChainPolicy BuildChainPolicy(X509Certificate2 certificate)
-        {
-            // Now build the chain validation options.
-            var revocationFlag = Options.RevocationFlag;
-            var revocationMode = Options.RevocationMode;
-
-            if (certificate.IsSelfSigned())
-            {
-                // Turn off chain validation, because we have a self signed certificate.
-                revocationFlag = X509RevocationFlag.EntireChain;
-                revocationMode = X509RevocationMode.NoCheck;
-            }
-
-            var chainPolicy = new X509ChainPolicy
-            {
-                RevocationFlag = revocationFlag,
-                RevocationMode = revocationMode,
-            };
-
-            if (Options.ValidateCertificateUse)
-            {
-                chainPolicy.ApplicationPolicy.Add(ClientCertificateOid);
-            }
-
-            if (certificate.IsSelfSigned())
-            {
-                chainPolicy.VerificationFlags |= X509VerificationFlags.AllowUnknownCertificateAuthority;
-                chainPolicy.VerificationFlags |= X509VerificationFlags.IgnoreEndRevocationUnknown;
-                chainPolicy.ExtraStore.Add(certificate);
-            }
-
-            if (!Options.ValidateValidityPeriod)
-            {
-                chainPolicy.VerificationFlags |= X509VerificationFlags.IgnoreNotTimeValid;
-            }
-
-            return chainPolicy;
-        }
-
-        private AuthenticateResult Success(ClaimsPrincipal principal, X509Certificate2 certificate)
-        {
-            var props = new AuthenticationProperties
-            {
-                Items =
-                {
-                    { CertificateAuthenticationDefaults.CertificateItemsKey, certificate.GetRawCertDataString() },
-                },
-            };
-
-            var ticket = new AuthenticationTicket(principal, props, Scheme.Name);
-            return AuthenticateResult.Success(ticket);
         }
     }
 }
